seggregated domain specific and enclave specific funcitinality into their specific folder
